requirements.txt
ediri_vultr==2.20.1
paramiko==3.5.0
peewee==3.17.8
petname==2.6
pulumi==3.140.0
pycountry==24.6.1
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
Requests==2.32.3
streamlit==1.40.1
